fix pr comments.
