+++ pkg/filters/proxy/pool.go
@@ -35,7 +35,9 @@ import (
 	"github.com/megaease/easegress/pkg/resilience"
 	"github.com/megaease/easegress/pkg/tracing"
 	"github.com/megaease/easegress/pkg/util/fasttime"
+	"github.com/megaease/easegress/pkg/util/prometheushelper"
 	"github.com/megaease/easegress/pkg/util/readers"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
 // serverPoolError is the error returned by handler function of
@@ -215,7 +217,7 @@ func NewServerPool(proxy *Proxy, spec *ServerPoolSpec, name string) *ServerPool 
 		sp.failureCodes[code] = struct{}{}
 	}
 
-	sp.metrics = proxy.newMetrics(name)
+	sp.metrics = sp.newMetrics(name)
 	return sp
 }
 
@@ -593,4 +595,61 @@ func (sp *ServerPool) inFailureCodes(code int) bool {
 
 	_, exists := sp.failureCodes[code]
 	return exists
+}
+
+type (
+	// ProxyMetrics is the Prometheus Metrics of ProxyMetrics Object.
+	metrics struct {
+		TotalConnections      *prometheus.CounterVec
+		TotalErrorConnections *prometheus.CounterVec
+		RequestBodySize       prometheus.ObserverVec
+		ResponseBodySize      prometheus.ObserverVec
+	}
+)
+
+// newMetrics create the ProxyMetrics.
+func (sp *ServerPool) newMetrics(name string) *metrics {
+	commonLabels := prometheus.Labels{
+		"name":         name,
+		"kind":         Kind,
+		"clusterName":  sp.proxy.super.Options().ClusterName,
+		"clusterRole":  sp.proxy.super.Options().ClusterRole,
+		"instanceName": sp.proxy.super.Options().Name,
+	}
+	proxyLabels := []string{"clusterName", "clusterRole", "instanceName",
+		"name", "kind", "loadBalancePolicy", "filterPolicy"}
+	return &metrics{
+		TotalConnections: prometheushelper.NewCounter("proxy_total_connections",
+			"the total count of proxy connections",
+			proxyLabels).MustCurryWith(commonLabels),
+		TotalErrorConnections: prometheushelper.NewCounter("proxy_total_error_connections",
+			"the total count of proxy error connections",
+			proxyLabels).MustCurryWith(commonLabels),
+		RequestBodySize: prometheushelper.NewHistogram("proxy_request_body_size",
+			"a histogram of the total size of the request.",
+			proxyLabels).MustCurryWith(commonLabels),
+		ResponseBodySize: prometheushelper.NewHistogram(
+			"proxy_response_body_size",
+			"a histogram of the total size of the response.",
+			proxyLabels).MustCurryWith(commonLabels),
+	}
+}
+
+func (sp *ServerPool) exportPrometheusMetrics(stat *httpstat.Metric) {
+	labels := prometheus.Labels{
+		"loadBalancePolicy": "",
+		"filterPolicy":      "",
+	}
+	if sp.spec.LoadBalance != nil {
+		labels["loadBalancePolicy"] = sp.spec.LoadBalance.Policy
+	}
+	if sp.spec.Filter != nil {
+		labels["filterPolicy"] = sp.spec.Filter.Policy
+	}
+	sp.metrics.TotalConnections.With(labels).Inc()
+	if stat.StatusCode >= 400 {
+		sp.metrics.TotalErrorConnections.With(labels).Inc()
+	}
+	sp.metrics.RequestBodySize.With(labels).Observe(float64(stat.ReqSize))
+	sp.metrics.ResponseBodySize.With(labels).Observe(float64(stat.RespSize))
 }
