+++ pkg/object/httpserver/runtime.go
@@ -39,6 +39,8 @@ import (
 	"github.com/megaease/easegress/pkg/util/easemonitor"
 	"github.com/megaease/easegress/pkg/util/filterwriter"
 	"github.com/megaease/easegress/pkg/util/limitlistener"
+	"github.com/megaease/easegress/pkg/util/prometheushelper"
+	"github.com/prometheus/client_golang/prometheus"
 )
 
 const (
@@ -419,4 +421,67 @@ func (s *Status) ToMetrics(service string) []*easemonitor.Metrics {
 	}
 
 	return results
+}
+
+type (
+	metrics struct {
+		Health             *prometheus.GaugeVec
+		TotalRequests      *prometheus.CounterVec
+		TotalResponses     *prometheus.CounterVec
+		TotalErrorRequests *prometheus.CounterVec
+		RequestsDuration   prometheus.ObserverVec
+		RequestSizeBytes   prometheus.ObserverVec
+		ResponseSizeBytes  prometheus.ObserverVec
+	}
+)
+
+// newMetrics create the HttpServerMetrics.
+func (r *runtime) newMetrics(name string) *metrics {
+	commonLabels := prometheus.Labels{
+		"name":         name,
+		"kind":         Kind,
+		"clusterName":  r.superSpec.Super().Options().ClusterName,
+		"clusterRole":  r.superSpec.Super().Options().ClusterRole,
+		"instanceName": r.superSpec.Super().Options().Name,
+	}
+	httpserverLabels := []string{"clusterName", "clusterRole",
+		"instanceName", "name", "kind", "routerKind", "backend"}
+	return &metrics{
+		Health: prometheushelper.NewGauge(
+			"httpserver_health",
+			"show the status for the http server: 1 for ready, 0 for down",
+			httpserverLabels[:5]).MustCurryWith(commonLabels),
+		TotalRequests: prometheushelper.NewCounter(
+			"httpserver_total_requests",
+			"the total count of http requests",
+			httpserverLabels).MustCurryWith(commonLabels),
+		TotalResponses: prometheushelper.NewCounter(
+			"httpserver_total_responses",
+			"the total count of http responses",
+			httpserverLabels).MustCurryWith(commonLabels),
+		TotalErrorRequests: prometheushelper.NewCounter(
+			"httpserver_total_error_requests",
+			"the total count of http error requests",
+			httpserverLabels).MustCurryWith(commonLabels),
+		RequestsDuration: prometheushelper.NewHistogram(
+			"httpserver_requests_duration",
+			"request processing duration histogram",
+			httpserverLabels).MustCurryWith(commonLabels),
+		RequestSizeBytes: prometheushelper.NewHistogram(
+			"httpserver_requests_size_bytes",
+			"a histogram of the total size of the request. Includes body",
+			httpserverLabels).MustCurryWith(commonLabels),
+		ResponseSizeBytes: prometheushelper.NewHistogram(
+			"httpserver_responses_size_bytes",
+			"a histogram of the total size of the returned response body",
+			httpserverLabels).MustCurryWith(commonLabels),
+	}
+}
+
+func (r *runtime) exportState(state stateType) {
+	if state == stateRunning {
+		r.metrics.Health.WithLabelValues().Set(1)
+	} else {
+		r.metrics.Health.WithLabelValues().Set(0)
+	}
 }
